user acl and data limited,collection optimized,acl added(need more review),pagination added in users,url for admin changed

+++ app/Http/Controllers/API/BannerController.php
@@ -21,8 +21,10 @@ class BannerController extends Controller
 
     public function index()
     {
-        $banner = Banner::all();
-        return $banner;
+        if (\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
+            $banner = Banner::all();
+            return $banner;
+        }
     }
 
 
@@ -34,32 +36,33 @@ class BannerController extends Controller
 
         //converting the base64 image to unique name and with intercention to make Image
         //store it in img/banners folder
+        if (\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
+            if($request->img){
+                $name = time().'.' .explode('/', explode(':',substr($request->img,0,strpos($request->img,';')))[1])[1];
+                $img = \Image::make($request->img);
+                $img->resize(1920, null, function ($constraint) {
+                    $constraint->aspectRatio();
+                });
+                $img->save(public_path('img/banners/'.$name));
+                $request->merge(['img'=>$name]);
+            }
+            //validating the request
+            $this->validate($request,[
+                'title' =>'required|string|max:191',
+                'status'=>'required|integer|max:191',
+                'sub_title'=>'required|string|max:500',
+                'img'=>'required|string|max:191',
+            ]);
 
-        if($request->img){
-            $name = time().'.' .explode('/', explode(':',substr($request->img,0,strpos($request->img,';')))[1])[1];
-            $img = \Image::make($request->img);
-            $img->resize(1920, null, function ($constraint) {
-                $constraint->aspectRatio();
-            });
-            $img->save(public_path('img/banners/'.$name));
-            $request->merge(['img'=>$name]);
+            //creating the banner
+            Banner::create([
+                'title' => $request->input('title'),
+                'status' =>$request->input('status'),
+                'sub_title' =>$request->input('sub_title'),
+                'img' =>$request->input('img'),
+                'uploaded_by' =>  auth('api')->user()->id
+            ]);
         }
-        //validating the request
-        $this->validate($request,[
-            'title' =>'required|string|max:191',
-            'status'=>'required|integer|max:191',
-            'sub_title'=>'required|string|max:500',
-            'img'=>'required|string|max:191',
-        ]);
-
-        //creating the banner
-        Banner::create([
-            'title' => $request->input('title'),
-            'status' =>$request->input('status'),
-            'sub_title' =>$request->input('sub_title'),
-            'img' =>$request->input('img'),
-            'uploaded_by' =>  auth('api')->user()->id
-        ]);
     }
 
 
@@ -67,45 +70,49 @@ class BannerController extends Controller
 
     public function update(Request $request, $id)
     {
-        //finding the banner through id
-        $banner = Banner::findOrFail($id);
-        //validating the request
-        $this->validate($request,[
-            'title' =>'required|string|max:191',
-            'status'=>'required|integer|max:191',
-            'sub_title'=>'required|string|max:500',
-        ]);
+        if (\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
+            //finding the banner through id
+            $banner = Banner::findOrFail($id);
+            //validating the request
+            $this->validate($request,[
+                'title' =>'required|string|max:191',
+                'status'=>'required|integer|max:191',
+                'sub_title'=>'required|string|max:500',
+            ]);
 
-        //converting image
-        if($request->img != $banner->img){
-            $name = time().'.' .explode('/', explode(':',substr($request->img,0,strpos($request->img,';')))[1])[1];
-            $img = \Image::make($request->img);
-            $img->resize(1920, null, function ($constraint) {
-                $constraint->aspectRatio();
-            });
-            $img->save(public_path('img/banners/'.$name));
-            $request->merge(['img'=>$name]);
+            //converting image
+            if($request->img != $banner->img){
+                $name = time().'.' .explode('/', explode(':',substr($request->img,0,strpos($request->img,';')))[1])[1];
+                $img = \Image::make($request->img);
+                $img->resize(1920, null, function ($constraint) {
+                    $constraint->aspectRatio();
+                });
+                $img->save(public_path('img/banners/'.$name));
+                $request->merge(['img'=>$name]);
+            }
+            //updating the banner
+            $banner->update([
+                'title' => $request->input('title'),
+                'status' =>$request->input('status'),
+                'sub_title' =>$request->input('sub_title'),
+                'img' =>$request->input('img'),
+            ]);
         }
-        //updating the banner
-        $banner->update([
-            'title' => $request->input('title'),
-            'status' =>$request->input('status'),
-            'sub_title' =>$request->input('sub_title'),
-            'img' =>$request->input('img'),
-        ]);
     }
 
     public function destroy($id)
     {
-        //finding the banner through id
-        $banner = Banner::findOrFail($id);
-        //searching for old image
-        $image = public_path('img/banners/').$banner->img;
-        //if file exists deleting it
-        if(file_exists($image)){
-            @unlink($image);
+        if (\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
+            //finding the banner through id
+            $banner = Banner::findOrFail($id);
+            //searching for old image
+            $image = public_path('img/banners/').$banner->img;
+            //if file exists deleting it
+            if(file_exists($image)){
+                @unlink($image);
+            }
+            //finally deleting the banner
+            $banner->delete();
         }
-        //finally deleting the banner
-        $banner->delete();
     }
 }
