+++ app/Http/Controllers/API/CategoryController.php
@@ -19,8 +19,10 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        if (\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
         $categories = Category::all();
         return $categories;
+        }
     }
 
     /**
@@ -41,26 +43,28 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        if (\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
         //validating the request
-        $this->validate($request,[
-            'name' =>'required|string|max:191',
-            'image'=>'required',
-        ]);
+                $this->validate($request,[
+                    'name' =>'required|string|max:191',
+                    'image'=>'required',
+                ]);
 
-           if($request->image){
-            $name = time().'.' .explode('/', explode(':',substr($request->image,0,strpos($request->image,';')))[1])[1];
-            $img = \Image::make($request->image);
-            $img->resize(600, null, function ($constraint) {
-                $constraint->aspectRatio();
-            });
-            $img->save(public_path('img/category/'.$name));
-            $request->merge(['image'=>$name]);
-        }
-        Category::create([
-            'name'=>$request->input('name'),
-            'image'=>$request->input('image'),
-            'user_id'=>auth('api')->user()->id
-        ]);
+                   if($request->image){
+                    $name = time().'.' .explode('/', explode(':',substr($request->image,0,strpos($request->image,';')))[1])[1];
+                    $img = \Image::make($request->image);
+                    $img->resize(600, null, function ($constraint) {
+                        $constraint->aspectRatio();
+                    });
+                    $img->save(public_path('img/category/'.$name));
+                    $request->merge(['image'=>$name]);
+                }
+                Category::create([
+                    'name'=>$request->input('name'),
+                    'image'=>$request->input('image'),
+                    'user_id'=>auth('api')->user()->id
+                ]);
+            }
 
     }
 
@@ -95,24 +99,26 @@ class CategoryController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $category = Category::findOrFail($id);
-        $this->validate($request,[
-            'name' =>'required|string|max:191',
-        ]);
-         if($request->image != $category->image){
-            $name = time().'.' .explode('/', explode(':',substr($request->image,0,strpos($request->image,';')))[1])[1];
-            $img = \Image::make($request->image);
-            $img->resize(600, null, function ($constraint) {
-                $constraint->aspectRatio();
-            });
-            $img->save(public_path('img/category/'.$name));
-            $request->merge(['image'=>$name]);
+        if (\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
+            $category = Category::findOrFail($id);
+            $this->validate($request,[
+                'name' =>'required|string|max:191',
+            ]);
+             if($request->image != $category->image){
+                $name = time().'.' .explode('/', explode(':',substr($request->image,0,strpos($request->image,';')))[1])[1];
+                $img = \Image::make($request->image);
+                $img->resize(600, null, function ($constraint) {
+                    $constraint->aspectRatio();
+                });
+                $img->save(public_path('img/category/'.$name));
+                $request->merge(['image'=>$name]);
+            }
+            $category->update([
+                'name'=>$request->input('name'),
+                'image'=>$request->input('image'),
+                'user_id'=>auth('api')->user()->id
+            ]);
         }
-        $category->update([
-            'name'=>$request->input('name'),
-            'image'=>$request->input('image'),
-            'user_id'=>auth('api')->user()->id
-        ]);
     }
 
     /**
@@ -123,14 +129,16 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
-         $category = Category::findOrFail($id);
-        //find the image
-        $image = public_path('img/category/').$category->image;
-        //delete the image
-        if(file_exists($image)){
-            @unlink($image);
+        if (\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
+             $category = Category::findOrFail($id);
+            //find the image
+            $image = public_path('img/category/').$category->image;
+            //delete the image
+            if(file_exists($image)){
+                @unlink($image);
+            }
+            //delete the gallery
+            $category->delete();
         }
-        //delete the gallery
-        $category->delete();
     }
 }
